feat: modularize core, combat, ai systems into structured modules
